refactor(home): Refactor state management by centralizing Cubits in HomeScreen
This commit refactors the state management on the home screen. It centralizes the initialization and provision of all home-related Cubits (Categories, Daily Offers, Suggestions, etc.) within the `HomeScreen`'s state.

This change eliminates the need for individual `BlocProvider` wrappers in each home screen section widget, simplifying the widget tree. A `MultiBlocProvider` is now used at the top level of the `HomeScreen` to provide these cubits down the tree.

Additionally, a pull-to-refresh functionality has been implemented to reload all data for the home screen widgets concurrently. A new `LifecycleEventHandler` is also introduced to refresh data when the app resumes.